fix for colliding and update on UI

+++ HAPI_Start/HAPI_Start/Bullet.cpp
@@ -6,7 +6,7 @@ Bullet::Bullet()
 {
 }
 
-void Bullet::Update(float dt) {
+void Bullet::Update(long dt) {
 
 	if(this->m_bDirection == LEFT) this->SetPosition(this->m_v2Pos.x - dt, this->m_v2Pos.y);
 	if (this->m_bDirection == RIGHT) this->SetPosition(this->m_v2Pos.x + dt, this->m_v2Pos.y);
